changed type of id

+++ src/main/java/com/halmaks/coins/controller/CustomerController.java
@@ -33,7 +33,7 @@ public class CustomerController {
     }
 
     @GetMapping("/customer/{id}")
-    public ResponseEntity<Customer> getCustomerById(@PathVariable(value = "id") Long customerId)
+    public ResponseEntity<Customer> getCustomerById(@PathVariable(value = "id") Integer customerId)
         throws ResourceNotFoundException {
         Customer customer = customerRepository.findById(customerId)
           .orElseThrow(() -> new ResourceNotFoundException("Customer not found for this id :: " + customerId));
@@ -46,7 +46,7 @@ public class CustomerController {
     }
 
     @PutMapping("/customer/{id}")
-    public ResponseEntity<Customer> updateCustomer(@PathVariable(value = "id") Long customerId,
+    public ResponseEntity<Customer> updateCustomer(@PathVariable(value = "id") Integer customerId,
          @Valid @RequestBody Customer customerDetails) throws ResourceNotFoundException {
         Customer customer = customerRepository.findById(customerId)
         .orElseThrow(() -> new ResourceNotFoundException("Customer not found for this id :: " + customerId));
@@ -59,7 +59,7 @@ public class CustomerController {
     }
 
     @DeleteMapping("/customer/{id}")
-    public Map<String, Boolean> deleteCustomer(@PathVariable(value = "id") Long customerId)
+    public Map<String, Boolean> deleteCustomer(@PathVariable(value = "id") Integer customerId)
          throws ResourceNotFoundException {
         Customer customer = customerRepository.findById(customerId)
        .orElseThrow(() -> new ResourceNotFoundException("Customer not found for this id :: " + customerId));
